Refactored the code and improved the output format
chiefly, I removed a lot of duplicate code and improved the output format to print out a table (here's a [screenshot](https://imgur.com/DT4enHV)), which allowed for much more compact and easy-to-scan output. This was only done for the data printouts, so the "personality" is preserved.

The refactored code is heavily stateful, like the original. The state is now encapsulated within a class, which while not ideal is better than having it scattered throughout the function, and several inefficient constructs (like that "yes" variable) have been simplified.